// src/lib/types.ts
// export type fontWeights =
//   | '100'
//   | '200'
//   | '300'
//   | '400'
//   | '500'
//   | '600'
//   | '700'
//   | '800'
//   | '900'

// export type DisplayType =
//   | 'inline'
//   | 'block'
//   | 'inline-block'
//   | 'flex'
//   | 'inline-flex'
//   | 'grid'
//   | 'inline-grid'

export type textSize =
  | 'xs'
  | 'sm'
  | 'base'
  | 'lg'
  | 'xl'
  | '2xl'
  | '3xl'
  | '4xl'
  | '5xl'
  | '6xl'
  | '7xl'
  | '8xl'
  | '9xl';

export type textAlign =
  | 'left'
  | 'right'
  | 'center'
  | 'justify'
  | 'start'
  | 'end';

export type textStyle = 'italic' | 'not-italic';

export type textVariantNumeric =
  | 'normal'
  | 'ordinal'
  | 'slashed'
  | 'lining'
  | 'oldstyle'
  | 'proportional'
  | 'tabular'
  | 'diagonal'
  | 'stacked';

export type textSpacing =
  | 'tighter'
  | 'tight'
  | 'normal'
  | 'wide'
  | 'wider'
  | 'widest';

export type textIndent =
  | '0'
  | 'px'
  | '0.5'
  | '1'
  | '1.5'
  | '2'
  | '2.5'
  | '3'
  | '3.5'
  | '4'
  | '5'
  | '6'
  | '7'
  | '8'
  | '9'
  | '10'
  | '11'
  | '12'
  | '14'
  | '16'
  | '20'
  | '24'
  | '28'
  | '32'
  | '36'
  | '40'
  | '44'
  | '48'
  | '52'
  | '56'
  | '60'
  | '64'
  | '72'
  | '80'
  | '96';

export type textBreak = 'normal' | 'words' | 'all' | 'keep';

// export type objectHeightAndWidth =
//   '0'
//   | "px"
//   | '0.5'
//   | '1'
//   | '1.5'
//   | '2'
//   | '2.5'
//   | '3'
//   | '3.5'
//   | '4'
//   | '5'
//   | '6'
//   | '7'
//   | '8'
//   | '9'
//   | '10'
//   | '11'
//   | '12'
//   | '14'
//   | '16'
//   | '20'
//   | '24'
//   | '28'
//   | '32'
//   | '36'
//   | '40'
//   | '44'
//   | '48'
//   | '52'
//   | '56'
//   | '60'
//   | '64'
//   | '72'
//   | '80'
//   | '96'
//   | '1/2'
//   | '1/3'
//   | '1/4'
//   | '2/4'
//   | '3/4'
//   | '1/5'
//   | '2/5'
//   | '3/5'
//   | '4/5'
//   | '1/6'
//   | '2/6'
//   | '3/6'
//   | '4/6'
//   | '5/6'
//   | 'auto'
//   | 'fit'
//   | 'full'
//   | 'screen'